Move type classes into their own package
