+++ common/src/main/scala/uk/ac/wellcome/storage/VersionedHybridStore.scala
@@ -5,9 +5,10 @@ import com.gu.scanamo.DynamoFormat
 import io.circe.{Decoder, Encoder}
 import shapeless.ops.hlist.{Collect, LeftFolder, Prepend, Zip}
 import shapeless.{Id => ShapelessId, _}
-import uk.ac.wellcome.dynamo.{IdGetter, VersionGetter, VersionedDao}
+import uk.ac.wellcome.dynamo.VersionedDao
 import uk.ac.wellcome.models._
 import uk.ac.wellcome.s3.S3ObjectStore
+import uk.ac.wellcome.type_classes.{IdGetter, VersionGetter, VersionUpdater}
 import uk.ac.wellcome.utils.GlobalExecutionContext._
 
 import scala.annotation.Annotation
