+++ common/src/test/scala/uk/ac/wellcome/dynamo/VersionedDaoTest.scala
@@ -1,12 +1,7 @@
 package uk.ac.wellcome.dynamo
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
-import com.amazonaws.services.dynamodbv2.model.{
-  AttributeValue,
-  ConditionalCheckFailedException,
-  GetItemRequest,
-  PutItemRequest
-}
+import com.amazonaws.services.dynamodbv2.model.{AttributeValue, ConditionalCheckFailedException, GetItemRequest, PutItemRequest}
 import com.gu.scanamo.error.DynamoReadError
 import com.gu.scanamo.syntax._
 import com.gu.scanamo.{DynamoFormat, Scanamo}
@@ -17,8 +12,9 @@ import org.scalatest.mockito.MockitoSugar
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.locals.DynamoDBLocal
 import uk.ac.wellcome.models.aws.DynamoConfig
-import uk.ac.wellcome.models.{Id, VersionUpdater, Versioned}
+import uk.ac.wellcome.models.{Id, Versioned}
 import uk.ac.wellcome.test.utils.ExtendedPatience
+import uk.ac.wellcome.type_classes.VersionUpdater
 
 case class TestVersioned(id: String, data: String, version: Int)
     extends Versioned
